Split registry up and two separate packages
