+++ xds/internal/xdsclient/xdsresource/tests/unmarshal_cds_test.go
@@ -34,6 +34,7 @@ import (
 	"google.golang.org/grpc/serviceconfig"
 	"google.golang.org/grpc/xds/internal/balancer/ringhash"
 	"google.golang.org/grpc/xds/internal/balancer/wrrlocality"
+	_ "google.golang.org/grpc/xds/internal/xdsclient/xdslbregistry/converter" // To register the xDS LB converters.
 	"google.golang.org/grpc/xds/internal/xdsclient/xdsresource"
 	"google.golang.org/protobuf/types/known/wrapperspb"
